refactor(file-storage): merge quota + usage adapters into external_clients
Both `infra/quota.rs` and `infra/usage.rs` were thin optional external-service
client-adapter modules of identical shape (a trait + a small decision/delta type
for a cloud integration). `FileService` imported both, spending two fan-out edges
on one concern. Fold them into a single `infra/external_clients.rs` so the hub
depends on one module instead of two (the HK "fold same-concern edges behind one
facade" remedy).

service.rs HK 2,019,780 -> 1,740,024 (fan_out 14 -> 13, -14%). No behaviour
change; 165 tests pass, clippy + fmt clean.
