+++ gears/file-storage/file-storage/src/infra/mod.rs
@@ -4,11 +4,11 @@
 //! - `content` — hashing, magic-byte mime validation, Range parsing
 //! - `backend` — pluggable storage-backend abstraction (local-fs, in-memory)
 //! - `signed_url` — Ed25519-signed opaque content tokens (issuer + verifier)
+//! - `external_clients` — optional external-service adapters (quota, usage reporting)
 
 pub mod authz;
 pub mod backend;
 pub mod content;
-pub mod quota;
+pub mod external_clients;
 pub mod signed_url;
 pub mod storage;
-pub mod usage;
